MAINT: Fix deprecated non-integer index to numpy arrays

+++ scipy/stats/stats.py
@@ -1470,15 +1470,15 @@ def scoreatpercentile(a, per, limit=(), interpolation_method='fraction'):
 
     idx = per /100. * (values.shape[0] - 1)
     if (idx % 1 == 0):
-        score = values[idx]
+        score = values[int(idx)]
     else:
         if interpolation_method == 'fraction':
             score = _interpolate(values[int(idx)], values[int(idx) + 1],
                                  idx % 1)
         elif interpolation_method == 'lower':
-            score = values[np.floor(idx)]
+            score = values[int(np.floor(idx))]
         elif interpolation_method == 'higher':
-            score = values[np.ceil(idx)]
+            score = values[int(np.ceil(idx))]
         else:
             raise ValueError("interpolation_method can only be 'fraction', " \
                              "'lower' or 'higher'")
